introduced separate class for the text UI;\n also, minor text fixes
