+++ railroad/cargo_train.rb
@@ -3,11 +3,10 @@ require_relative 'cargo_wagon'
 
 class CargoTrain < Train
   def add_wagon(wagon)
-    if wagon.class == CargoWagon
+    if wagon.instance_of?(CargoWagon)
       super(wagon)
     else
-      puts "Alarm! Wrong wagon type: CargoTrain needs CargoWagon"
+      puts 'Alarm! Wrong wagon type: CargoTrain needs CargoWagon'
     end
   end
 end
-
