+++ railroad/passenger_train.rb
@@ -3,11 +3,10 @@ require_relative 'passenger_wagon'
 
 class PassengerTrain < Train
   def add_wagon(wagon)
-    if wagon.class == PassengerWagon
+    if wagon.instance_of?(PassengerWagon)
       super(wagon)
     else
-      puts "Alarm! Wrong wagon type: PassengerTrain needs PassengerWagon"
+      puts 'Alarm! Wrong wagon type: PassengerTrain needs PassengerWagon'
     end
   end
 end
-
