+++ railroad/passenger_wagon.rb
@@ -1,4 +1,3 @@
 require_relative 'wagon'
 
 class PassengerWagon < Wagon; end
-
