+++ railroad/station.rb
@@ -20,4 +20,3 @@ class Station
     @trains.delete train
   end
 end
-
