+++ railroad/train.rb
@@ -22,7 +22,7 @@ class Train
   def remove_wagon(wagon)
     @wagon_list << wagon if speed.zero?
   end
-  
+
   def remove_last_wagon
     @wagon_list.pop
   end
