+++ railroad/wagon.rb
@@ -1,2 +1,1 @@
 class Wagon; end
-
